requirements.txt
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
pydeck==0.7.1
PyYAML==6.0
streamlit==1.10.0
